Remove exception wrapping
When libatspi throws an exception, return the exception as translated by
pygobject. Previously we were raising LookupError since that was
generally the error raised by AT-SPI 1 in similar cases, but this was
kludgy and raised the same exception regardless of the original error,
so an AT could not distinguish between various types of exceptions.
This means that consumers of pyatspi now need to test for pygobject
exceptions (gi._glib.GError, which is a sub-class of RuntimeError).