Handle match rules with multiple attribute values associated with a key
Specifying multiple attribute strings with a given key, with the
intention of accepting any of a specified set of values, would result in
the last specified string overriding previous ones. Instead, delimit
values with a :, and also allow a : in a string to be escaped with a \.
(Another approach would have been to have libatspi allow creation of a
match rule from a list, rather than a hash, but trying to avoid an API
change, especially when we're in API freeze).

https://bugzilla.gnome.org/show_bug.cgi?id=700865