Rename enum module and Enum class not to conflict with stdlib
Python 3 stdlib has an enum module with an Enum class...and so
do we. They do not work at all the same. This is a bad thing,
and causes modules which rely on the Python stdlib version to
blow up when they're used with pyatspi in the Python path, like
re:

https://bugzilla.redhat.com/show_bug.cgi?id=1406946

To avoid this, let's rename our 'enum' module to 'atspienum'
and our 'Enum' class to 'AtspiEnum'. This is an API change, but
it's rather unavoidable, we just can't keep the name 'enum' and
use regexes on Python 3.6. AFAICS, nothing besides pyatspi is
actually using this module / class.