Fix creating dom nodes
Python 3.13 introduced a fix
https://github.com/python/cpython/pull/142146

which brought a regression
https://github.com/python/cpython/issues/142754

Discussion shows that one is not supposed to use Element to create
nodes. It used to work but is not really supported. They fixed the
issue in
https://github.com/python/cpython/pull/142794/files

but better create dom nodes the proper way with doc.createElement.

collectiontest.py didn't actually need to create elements any more.